memory leak problem
When I run the traffic generator, I find that VIRT is increasing with time going by, even up to 70GB. So I check the source code and find there is a memory leak problem in server.c file.  The issue is that there is no thread release in while function, so the number of thread is growing when a new request comes continuously, which caused the huge VIRT occupied, so I fix this problem by adding pthread_detach function as shown in line 128.